Added check for 64 bit portability issues in Visual Studio. Plus fixed some warnings.

+++ src/types/schema/schema.cpp
@@ -1823,7 +1823,7 @@ void Schema::serialize(::zorba::serialization::Archiver &ar)
        try
        {
          theGrammarPool->serializeGrammars(&binmemoutputstream);
-         size = binmemoutputstream.getSize();
+         size = (unsigned int)binmemoutputstream.getSize();
          binchars = (unsigned char*)binmemoutputstream.getRawBuffer();
        }
        catch (...)
